requirements.txt
Django==3.2.4
djangorestframework==3.13
psycopg2
django-cors-headers==3.11.0